Remove old landing page.
