--- src/v2/Apps/Consign/Routes/MarketingLanding/Components/ConsignMeta.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import { MetaTags } from "v2/Components/MetaTags"
-
-export const ConsignMeta: React.FC = () => {
-  return (
-    <MetaTags
-      pathname="/consign"
-      title="Sell Artwork with Artsy | Art Consignment | Artsy"
-      description="Sell and consign your art with Artsy. Get competitive offers from the world’s top auction houses and galleries to sell art from your collection. Submit today at no cost."
-    />
-  )
-}
