+++ src/v2/Apps/Consign/Routes/MarketingLanding/Components/FAQ.tsx
@@ -1,14 +1,16 @@
 import { Expandable, Join, Spacer, Text } from "@artsy/palette"
 import { AnalyticsSchema, useTracking } from "v2/System"
-import { useRouter } from "v2/System/Router/useRouter"
 import { ContextModule, OwnerType } from "@artsy/cohesion"
 
-export const FAQ: React.FC = () => {
+export interface FAQProps {
+  shouldTrackClickEvent?: boolean
+}
+
+export const FAQ: React.FC<FAQProps> = ({ shouldTrackClickEvent }) => {
   const { trackEvent } = useTracking()
-  const { match } = useRouter()
 
   const trackClickedFAQ = (label: string) => {
-    if (match?.params?.id) {
+    if (shouldTrackClickEvent) {
       trackEvent({
         action_type: AnalyticsSchema.ActionType.ClickedFAQ,
         context_module: ContextModule.consignSubmissionFlow,
@@ -23,7 +25,7 @@ export const FAQ: React.FC = () => {
       label: "What does it cost to sell with Artsy?",
       value: (
         <TextItem>
-          We offer no upfront fees. When the artwork sells, we charge a nominal
+          There are no upfront fees. When the artwork sells, we charge a nominal
           seller’s commission as a percentage of the sale price. The commission
           is determined based on the total value of works you sell with us. You
           won’t be asked to cover any additional costs, such as shipping,
@@ -32,40 +34,13 @@ export const FAQ: React.FC = () => {
       ),
     },
     {
-      label: "What happens once I submit an artwork?",
+      label: "How long does it take to sell with Artsy?",
       value: (
         <TextItem>
-          Based on our proprietary market data, we will let you know within 48
-          hours whether your artwork has sufficient demand within the secondary
-          market. If so, your work will be approved. Once approved, you will
-          receive an offer from Artsy to include your work in an upcoming online
-          auction. The offer will have an estimated artwork range and auction
-          details. If you are interested in selling your work based on the
-          offer, you can accept it or ask further questions to our specialists,
-          who will guide you through the process.
-        </TextItem>
-      ),
-    },
-    {
-      label:
-        "How is selling with Artsy different from consigning with a traditional auction house?",
-      value: (
-        <TextItem>
-          Unlike a consignment with a traditional auction house, Artsy does not
-          require any upfront costs of shipping, insurance, or marketing. This
-          means no upfront fees for you. This also allows you to sell your work
-          in an online auction faster, as we hold multiple auctions a month. In
-          case your work doesn’t find a buyer right away, the auction results
-          won’t be published and the work won’t get “burned.”
-        </TextItem>
-      ),
-    },
-    {
-      label: "How long does it take to sell art with Artsy?",
-      value: (
-        <TextItem>
-          We hold multiple themed auctions per month, which means that your
-          artwork can be sold within 2–6 weeks.
+          We hold multiple themed auctions per month, which means your artwork
+          can be sold at auction within 2–6 weeks. Timing of private sales or
+          direct listing on Artsy can vary; chat with one of our specialists to
+          learn more.
         </TextItem>
       ),
     },
@@ -74,70 +49,67 @@ export const FAQ: React.FC = () => {
       value: (
         <TextItem>
           <p>
-            At this time, Artsy does not partner directly with artists. Instead,
-            we offer partnerships for galleries, museums, and other
-            institutions. If you work with a gallery, we ask you to get in touch
-            with us at{" "}
+            Currently, Artsy does not directly partner with artists outside of
+            our approved artists list. Instead, we offer partnerships for
+            galleries, museums, and other institutions.
+          </p>
+          <p>
+            If you are represented by a gallery, you can encourage them to
+            partner with us by getting in touch at{" "}
             <a href="mailto:partners@artsy.net">partners@artsy.net</a>.
-            Otherwise, we recommend reaching out to your gallery and asking them
-            to join Artsy by visiting{" "}
-            <a href="https://partners.artsy.net" target="_blank">
-              partners.artsy.net
+          </p>
+          <p>
+            If you’re looking for gallery representation, you might be
+            interested in reading Artsy Editorial’s{" "}
+            <a
+              href="/article/artsy-editorial-artists-gallery-representation"
+              target="_blank"
+            >
+              Gallery Representation Guide
+            </a>{" "}
+            and{" "}
+            <a href="/series/working-artists-guide" target="_blank">
+              Working Artist’s Guide
             </a>
             .
           </p>
-          <p>
-            Artsy has a{" "}
-            <a href="/consign" target="_blank">
-              Sell with Artsy program
-            </a>{" "}
-            in which we currently accept works by artists that are already in
-            our database who have a resale market and substantial collector
-            demand. To learn more about our Sell with Artsy program, please
-            visit our Help Center article on{" "}
-            <a
-              href="https://support.artsy.net/hc/en-us/articles/1500006366381"
-              target="_blank"
-            >
-              what do we look for in consignment submissions
-            </a>
-            ? Please note that not being accepted for our Sell with Artsy
-            program is in no way a reflection or judgement of artistic ability,
-            potential now or in the future, or commitment to the discipline; it
-            is simply a result of the assessment of the information that we were
-            given at the time and possible services we are able to currently
-            provide.
-          </p>
-          <p>
-            We'd also like to bring to your attention{" "}
-            <a href="/series/working-artists-guide" target="_blank">
-              The Working Artist’s Guide
-            </a>{" "}
-            which is a resource for artists put together by our Editorial team.
-            We hope it can assist you and that you continue to enjoy exploring
-            Artsy.
-          </p>
         </TextItem>
       ),
     },
     {
-      label: "Can I edit my submission?",
+      label: "I am a gallerist. How can I partner with Artsy?",
       value: (
         <TextItem>
-          To edit your existing submissions or add further details, please
-          contact sell@artsy.net with the subject line “Edit my submission” and
-          one of our specialists will assist you. In the body of your email,
-          please include the submission ID number, which can be found in the
-          email receipt you received from Artsy.
+          Visit{" "}
+          <a href="https://partners.artsy.net" target="_blank">
+            partners.artsy.net
+          </a>{" "}
+          to apply to become a gallery partner and learn more about our
+          marketplace plans.
         </TextItem>
       ),
     },
     {
-      label: "How do I contact an Artsy specialist?",
+      label: "Which items do you accept?",
       value: (
         <TextItem>
-          You can contact an Artsy specialist at any time by emailing
-          sell@artsy.net.
+          We are currently accepting unique and limited-edition works of art by
+          modern, contemporary, and emerging artists who have collector demand
+          on Artsy. Our in-house experts cover Post-War and Contemporary Art;
+          Prints and Multiples; and Street Art and Photographs.
+        </TextItem>
+      ),
+    },
+    {
+      label: "How do you price artworks for sale?",
+      value: (
+        <TextItem>
+          To assess a work’s value, our team uses publicly available auction
+          results, as well as Artsy’s own database of primary market sales. We
+          also take into account market trends (including both an individual
+          artist’s market and macro trends in the art market), artwork
+          condition, and sales channel, as well as other factors that may impact
+          a work’s value, such as medium, provenance, subject matter, or period.
         </TextItem>
       ),
     },
@@ -145,7 +117,7 @@ export const FAQ: React.FC = () => {
 
   return (
     <>
-      <Text mb={4} variant="xl" id="jump--faq">
+      <Text mb={4} variant="lg" id="jump--faq">
         Frequently Asked Questions
       </Text>
       <Join separator={<Spacer mt={4} />}>
@@ -161,7 +133,7 @@ export const FAQ: React.FC = () => {
 
 const TextItem: React.FC = ({ children }) => {
   return (
-    <Text variant="sm" color="black60" pr={2}>
+    <Text variant="md" color="black60" pr={2}>
       {children}
     </Text>
   )
