--- src/v2/Apps/Consign/Routes/MarketingLanding/Components/InDemandNow/ConsignInDemandNow.tsx
@@ -1,517 +1,0 @@
-import * as React from "react";
-import {
-  Text,
-  BlueChipIcon,
-  Box,
-  Column,
-  Flex,
-  GridColumns,
-  Image,
-  Spacer,
-  ResponsiveBox,
-  FullBleed,
-} from "@artsy/palette"
-import { SystemQueryRenderer } from "v2/System/Relay/SystemQueryRenderer"
-import { useSystemContext } from "v2/System"
-import { graphql } from "react-relay"
-import { ConsignInDemandNowQuery } from "v2/__generated__/ConsignInDemandNowQuery.graphql"
-import { ConsignTopArtists } from "./ConsignTopArtists"
-import { DemandRank } from "./DemandRank"
-import { formatCentsToDollars } from "v2/Apps/Consign/Routes/MarketingLanding/Utils/formatCentsToDollars"
-import { Media } from "v2/Utils/Responsive"
-import { AppContainer } from "v2/Apps/Components/AppContainer"
-import { HorizontalPadding } from "v2/Apps/Components/HorizontalPadding"
-
-export const ConsignInDemandNow = () => {
-  return (
-    <FullBleed bg="black100" color="white100">
-      <AppContainer>
-        <HorizontalPadding py={[2, 4]}>
-          <Text variant="xl" mb={2} pb={2} borderBottom="1px solid">
-            In Demand Now
-          </Text>
-
-          <Spacer my={2} />
-
-          <ConsignInDemandNowQueryRenderer />
-
-          <Spacer my={4} />
-
-          <ConsignTopArtists />
-        </HorizontalPadding>
-      </AppContainer>
-    </FullBleed>
-  )
-}
-
-const ConsignInDemandNowQueryRenderer: React.FC = () => {
-  const { relayEnvironment } = useSystemContext()
-
-  return (
-    <SystemQueryRenderer<ConsignInDemandNowQuery>
-      environment={relayEnvironment}
-      variables={{
-        artistInternalId: "4ed901b755a41e0001000a9f", // Kehinde Wiley
-        artistSlug: "kehinde-wiley",
-        medium: "PAINTING",
-      }}
-      // FIXME: Must be logged in for diffusion data to work
-      query={graphql`
-        query ConsignInDemandNowQuery(
-          $artistInternalId: ID!
-          $artistSlug: String!
-          $medium: String!
-        ) {
-          artist(id: $artistSlug) {
-            birthday
-            nationality
-            auctionResultsConnection(first: 1, sort: DATE_DESC) {
-              edges {
-                node {
-                  internalID
-                  title
-                  dimensionText
-                  images {
-                    larger {
-                      url
-                      resized {
-                        srcSet
-                      }
-                    }
-                  }
-                  description
-                  dateText
-                  organization
-                  saleDate
-                  priceRealized {
-                    display
-                    centsUSD
-                  }
-                }
-              }
-            }
-          }
-          marketPriceInsights(artistId: $artistInternalId, medium: $medium) {
-            annualLotsSold
-            annualValueSoldCents
-            artistId
-            artistName
-            artsyQInventory
-            createdAt
-            demandRank
-            demandTrend
-            highRangeCents
-            largeHighRangeCents
-            largeLowRangeCents
-            largeMidRangeCents
-            liquidityRank
-            lowRangeCents
-            medianSaleToEstimateRatio
-            medium
-            mediumHighRangeCents
-            mediumLowRangeCents
-            mediumMidRangeCents
-            midRangeCents
-            sellThroughRate
-            smallHighRangeCents
-            smallLowRangeCents
-            smallMidRangeCents
-            updatedAt
-          }
-        }
-      `}
-      render={({ props, error }) => {
-        // FIXME: Error handling
-        if (error) {
-          return null
-        }
-
-        return (
-          <>
-            <Media greaterThanOrEqual="md">
-              <InDemandNowLarge {...props!} />
-            </Media>
-
-            <Media lessThan="md">
-              <InDemandNowSmall {...props!} />
-            </Media>
-          </>
-        )
-      }}
-    />
-  )
-}
-
-const InDemandNowLarge: React.FC<
-  ConsignInDemandNowQuery["response"]
-> = props => {
-  const IMAGE_SIZE = 225
-
-  // FIXME: Add skeleton loading state
-  if (!props?.marketPriceInsights || !props?.artist) {
-    return null
-  }
-
-  const {
-    artistName,
-    averageSalePrice,
-    birthday,
-    demandRank,
-    highRangeDollars,
-    lastAuctionResult,
-    medianSaleToEstimateRatio,
-    nationality,
-    sellThroughRate,
-  } = computeProps(props)
-
-  return (
-    <>
-      <Text variant="xl" mb={1}>
-        {artistName}
-      </Text>
-
-      <Text variant="xs">
-        {nationality}, b. {birthday}
-      </Text>
-
-      <Flex>
-        <BlueChipIcon width={15} height={15} fill="white100" mr={0.5} />{" "}
-        <Text variant="xs">Blue-Chip Representation</Text>
-      </Flex>
-
-      <Spacer mt={4} />
-
-      <GridColumns gridRowGap={6}>
-        <Column span={[12, 12, 4]}>
-          {demandRank && (
-            <>
-              <DemandRank demandRank={demandRank} />
-
-              <Spacer mt={4} />
-            </>
-          )}
-
-          <Text variant="xs" textTransform="uppercase">
-            Highest Realized Price
-          </Text>
-
-          <Text variant="xl">{highRangeDollars}</Text>
-
-          {sellThroughRate && (
-            <>
-              <Spacer my={2} />
-
-              <Text variant="xs" textTransform="uppercase">
-                Sell-Through Rate
-              </Text>
-
-              <Text variant="xl">{sellThroughRate * 100}%</Text>
-            </>
-          )}
-
-          {medianSaleToEstimateRatio && (
-            <>
-              <Spacer my={2} />
-
-              <Text variant="xs" textTransform="uppercase">
-                Realized Price / Estimate
-              </Text>
-
-              {/* TODO: Follow-up on how real diffusion number (1.84) maps to designs (184%) */}
-              <Text variant="xl">{medianSaleToEstimateRatio * 100}%</Text>
-            </>
-          )}
-        </Column>
-
-        <Column span={[12, 12, 4]} textAlign="center">
-          <Box display="inline-block" mx="auto" textAlign="left">
-            <Flex flexDirection="column" maxWidth={IMAGE_SIZE}>
-              <Text variant="xs" textTransform="uppercase">
-                Recent Auction Result
-              </Text>
-
-              <Flex alignItems="baseline">
-                <Text variant="xl" mr={0.5}>
-                  $176,400
-                  {/* FIXME: This data is null in diffusion */}
-                  {/* {lastAuctionResult.priceRealized.display} */}
-                </Text>
-
-                <Text variant="xs" overflowEllipsis>
-                  Realized price
-                </Text>
-              </Flex>
-
-              {lastAuctionResult?.images?.larger && (
-                <>
-                  <Spacer my={1} />
-
-                  <Image
-                    src={lastAuctionResult.images.larger.url!}
-                    srcSet={lastAuctionResult.images.larger.resized?.srcSet!}
-                    alt={lastAuctionResult.title ?? ""}
-                    width={IMAGE_SIZE}
-                    height="auto"
-                  />
-                </>
-              )}
-
-              {lastAuctionResult && (
-                <>
-                  <Spacer my={1} />
-
-                  <Text variant="xs" overflowEllipsis>
-                    {lastAuctionResult.title}, {lastAuctionResult.dateText}
-                  </Text>
-
-                  <Text variant="xs">{lastAuctionResult.organization}</Text>
-                </>
-              )}
-            </Flex>
-          </Box>
-        </Column>
-
-        <Column span={[12, 12, 4]} textAlign="center">
-          <Box display="inline-block" textAlign="left" width={375}>
-            <Text variant="xs" textTransform="uppercase">
-              Average Sale Price, 2010–2020
-            </Text>
-
-            <Text variant="xl" mb={1}>
-              {averageSalePrice}
-            </Text>
-
-            <AverageSalePriceChart />
-          </Box>
-        </Column>
-      </GridColumns>
-    </>
-  )
-}
-
-const InDemandNowSmall: React.FC<
-  ConsignInDemandNowQuery["response"]
-> = props => {
-  const IMAGE_SIZE = 100
-
-  // FIXME: Add skeleton loading state
-  if (!props?.marketPriceInsights || !props?.artist) {
-    return null
-  }
-
-  const {
-    artistName,
-    averageSalePrice,
-    birthday,
-    demandRank,
-    highRangeDollars,
-    lastAuctionResult,
-    medianSaleToEstimateRatio,
-    nationality,
-    sellThroughRate,
-  } = computeProps(props)
-
-  return (
-    <>
-      <Text variant="xl" mb={1}>
-        {artistName}
-      </Text>
-
-      <Text variant="xs">
-        {nationality}, b. {birthday}
-      </Text>
-
-      <Flex>
-        <BlueChipIcon width={15} height={15} fill="white100" mr={0.5} />{" "}
-        <Text variant="xs">Blue-Chip Representation</Text>
-      </Flex>
-
-      <Spacer mt={2} />
-
-      <Flex justifyContent="space-between">
-        <Box>
-          <Box maxWidth={IMAGE_SIZE}>
-            {lastAuctionResult?.images?.larger && (
-              <>
-                <Image
-                  src={lastAuctionResult.images.larger.url!}
-                  srcSet={lastAuctionResult.images.larger.resized?.srcSet!}
-                  alt={lastAuctionResult.title ?? ""}
-                  width={IMAGE_SIZE}
-                  height="auto"
-                />
-
-                <Spacer my={0.5} />
-              </>
-            )}
-
-            {lastAuctionResult && (
-              <>
-                <Text variant="xs" overflowEllipsis>
-                  {lastAuctionResult.title}, {lastAuctionResult.dateText}
-                </Text>
-
-                <Text variant="xs">{lastAuctionResult.organization}</Text>
-              </>
-            )}
-          </Box>
-
-          <Spacer my={2} />
-
-          <Text variant="xs" textTransform="uppercase">
-            Realized Price
-          </Text>
-
-          <Text variant="xl">
-            $176,400
-            {/* FIXME: This data is null in diffusion */}
-            {/* {lastAuctionResult.priceRealized.display} */}
-          </Text>
-        </Box>
-
-        <Box>
-          <Text variant="xs" textTransform="uppercase">
-            Highest Realized Price
-          </Text>
-
-          <Text variant="xl">{highRangeDollars}</Text>
-
-          {sellThroughRate && (
-            <>
-              <Spacer my={1} />
-
-              <Text variant="xs" textTransform="uppercase">
-                Sell-Through Rate
-              </Text>
-
-              <Text variant="xl">{sellThroughRate * 100}%</Text>
-            </>
-          )}
-
-          {medianSaleToEstimateRatio && (
-            <>
-              <Spacer my={1} />
-
-              <Text variant="xs" textTransform="uppercase">
-                Realized Price/Estimate
-              </Text>
-
-              {/* TODO: Follow-up on how real diffusion number (1.84) maps to designs (184%) */}
-              <Text variant="xl">{medianSaleToEstimateRatio * 100}%</Text>
-            </>
-          )}
-        </Box>
-      </Flex>
-
-      <Spacer mt={4} />
-
-      {demandRank && (
-        <>
-          <DemandRank demandRank={demandRank} />
-
-          <Spacer mt={4} />
-        </>
-      )}
-
-      <Box textAlign="center">
-        <Box display="inline-block" textAlign="left" width="100%">
-          <Text variant="xs" textTransform="uppercase">
-            Average Sale Price, 2010–2020
-          </Text>
-
-          <Text variant="xl">{averageSalePrice}</Text>
-
-          <AverageSalePriceChart />
-        </Box>
-      </Box>
-    </>
-  )
-}
-
-const computeProps = ({
-  artist,
-  marketPriceInsights,
-}: ConsignInDemandNowQuery["response"]) => {
-  if (!artist || !marketPriceInsights) return {}
-
-  const {
-    artistName,
-    demandRank,
-    lowRangeCents,
-    highRangeCents,
-    medianSaleToEstimateRatio,
-    sellThroughRate,
-  } = marketPriceInsights
-
-  const { birthday, nationality, auctionResultsConnection } = artist
-  const lastAuctionResult = auctionResultsConnection?.edges?.[0]?.node
-
-  // TODO: Follow-up if this maps correctly to Highest Realized Price
-  const highRangeDollars = formatCentsToDollars(Number(highRangeCents))
-  const averageSalePrice = formatCentsToDollars(
-    (Number(lowRangeCents) + Number(highRangeCents)) / 2
-  )
-
-  return {
-    artistName,
-    averageSalePrice,
-    birthday,
-    demandRank,
-    highRangeDollars,
-    lastAuctionResult,
-    medianSaleToEstimateRatio,
-    nationality,
-    sellThroughRate,
-  }
-}
-
-// TODO: Extract to file hosted in Vanity
-const AverageSalePriceChart: React.FC = () => {
-  return (
-    <ResponsiveBox aspectWidth={375} aspectHeight={320} maxWidth={375}>
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        preserveAspectRatio="none"
-        height="320"
-        fill="none"
-        viewBox="0 0 375 320"
-        style={{ width: "100%", height: "100%" }}
-      >
-        <path
-          fill="#fff"
-          d="M325.616 145.643c0 1.974 1.162 3.08 3.108 3.22v1.274h.868v-1.288c1.82-.14 3.052-1.218 3.052-2.73 0-1.834-1.596-2.31-3.052-2.716v-3.402c.896.154 1.638.798 1.638 1.974h1.218c0-1.75-1.064-2.786-2.856-2.912v-1.148h-.868v1.148c-1.75.14-2.94 1.176-2.94 2.618 0 1.806 1.484 2.254 2.94 2.66v3.57c-1.204-.154-1.834-.924-1.834-2.268h-1.274zm5.698.574c0 .91-.644 1.554-1.722 1.694v-3.332c.966.294 1.722.7 1.722 1.638zm-4.2-4.676c0-.882.77-1.4 1.61-1.54v3.15c-.924-.294-1.61-.686-1.61-1.61zm8.552 7.182h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm6.794-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm8.326-1.064c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm-25.17-60.642c0 1.974 1.162 3.08 3.108 3.22v1.274h.868V82.58c1.82-.14 3.052-1.218 3.052-2.73 0-1.834-1.596-2.31-3.052-2.716v-3.402c.896.154 1.638.798 1.638 1.974h1.218c0-1.75-1.064-2.786-2.856-2.912v-1.148h-.868v1.148c-1.75.14-2.94 1.176-2.94 2.618 0 1.806 1.484 2.254 2.94 2.66v3.57c-1.204-.154-1.834-.924-1.834-2.268h-1.274zm5.698.574c0 .91-.644 1.554-1.722 1.694V78.31c.966.294 1.722.7 1.722 1.638zm-4.2-4.676c0-.882.77-1.4 1.61-1.54v3.15c-.924-.294-1.61-.686-1.61-1.61zm8.552 7.182h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm3.49-4.143l1.372.125c.266-.854.938-1.33 1.876-1.33 1.19 0 2.016.938 2.016 2.254 0 1.372-.826 2.24-2.072 2.24-1.078 0-1.848-.644-1.974-1.694h-1.344c.224 1.694 1.484 2.758 3.332 2.758 2.002 0 3.416-1.414 3.416-3.374 0-1.89-1.302-3.248-3.122-3.248-.952 0-1.554.308-2.226 1.008l.336-3.038h4.424v-1.12h-5.502l-.532 5.418zm11.233-5.629c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zM325.616 11.92c0 1.974 1.162 3.08 3.108 3.22v1.274h.868v-1.288c1.82-.14 3.052-1.218 3.052-2.73 0-1.834-1.596-2.31-3.052-2.716V6.278c.896.154 1.638.798 1.638 1.974h1.218c0-1.75-1.064-2.786-2.856-2.912V4.192h-.868V5.34c-1.75.14-2.94 1.176-2.94 2.618 0 1.806 1.484 2.254 2.94 2.66v3.57c-1.204-.154-1.834-.924-1.834-2.268h-1.274zm5.698.574c0 .91-.644 1.554-1.722 1.694v-3.332c.966.294 1.722.7 1.722 1.638zm-4.2-4.676c0-.882.77-1.4 1.61-1.54v3.15c-.924-.294-1.61-.686-1.61-1.61zM340.132 15v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422V15h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm8.326-1.064c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm-27.453 205.621c0 1.974 1.162 3.08 3.108 3.22v1.274h.868v-1.288c1.82-.14 3.052-1.218 3.052-2.73 0-1.834-1.596-2.31-3.052-2.716v-3.402c.896.154 1.638.798 1.638 1.974h1.218c0-1.75-1.064-2.786-2.856-2.912v-1.148h-.868v1.148c-1.75.14-2.94 1.176-2.94 2.618 0 1.806 1.484 2.254 2.94 2.66v3.57c-1.204-.154-1.834-.924-1.834-2.268h-1.274zm5.698.574c0 .91-.644 1.554-1.722 1.694v-3.332c.966.294 1.722.7 1.722 1.638zm-4.2-4.676c0-.882.77-1.4 1.61-1.54v3.15c-.924-.294-1.61-.686-1.61-1.61zm6.942 3.038l1.372.126c.266-.854.938-1.33 1.876-1.33 1.19 0 2.016.938 2.016 2.254 0 1.372-.826 2.24-2.072 2.24-1.078 0-1.848-.644-1.974-1.694h-1.344c.224 1.694 1.484 2.758 3.332 2.758 2.002 0 3.416-1.414 3.416-3.374 0-1.89-1.302-3.248-3.122-3.248-.952 0-1.554.308-2.226 1.008l.336-3.038h4.424v-1.12h-5.502l-.532 5.418zm11.234-5.628c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zM238.019 313.647v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm6.522 8.708h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm6.71.21c2.016 0 3.43-1.148 3.43-2.758 0-1.26-.63-2.058-2.002-2.38v-.07c1.134-.252 1.736-1.078 1.736-2.17 0-1.554-1.274-2.604-3.164-2.604-1.918 0-3.192 1.05-3.192 2.604 0 1.092.616 1.918 1.736 2.17v.07c-1.372.322-2.002 1.12-2.002 2.38 0 1.61 1.414 2.758 3.458 2.758zm0-1.036c-1.246 0-2.072-.77-2.072-1.806 0-1.078.826-1.778 2.072-1.778 1.232 0 2.058.7 2.058 1.778 0 1.036-.826 1.806-2.058 1.806zm0-4.662c-1.134 0-1.89-.63-1.89-1.624 0-.966.728-1.624 1.89-1.624 1.148 0 1.862.658 1.862 1.624 0 .994-.742 1.624-1.862 1.624zm39.413 5.488v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.61-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm10.989 8.708v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm-139.446 8.708v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm6.523 8.708h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm10-7.14c-.182-1.666-1.302-2.632-3.052-2.632-2.296 0-3.641 1.918-3.641 5.124 0 3.066 1.232 4.858 3.416 4.858 1.904 0 3.375-1.246 3.375-3.318 0-1.988-1.429-3.15-3.024-3.15-1.079 0-1.89.462-2.436 1.414v-.35c0-2.142.868-3.514 2.31-3.514.98 0 1.568.532 1.68 1.568h1.372zm-3.108 1.946c1.162 0 1.89.868 1.89 2.142 0 1.288-.77 2.212-1.89 2.212-1.162 0-2.03-.938-2.03-2.24 0-1.218.868-2.114 2.03-2.114zm-78.916 5.194v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.61-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm6.523 8.708h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm10.214-3.528h-1.722v-6.034h-1.498l-4.186 6.076v1.064h4.382v2.422h1.302v-2.422h1.722v-1.106zm-3.906-3.108c.378-.546.616-.91.924-1.47l.098.014c-.056.7-.07 2.002-.07 2.674v1.89h-3.08l2.128-3.108zm-77.257 6.636v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm6.523 8.708h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm9.565 0v-1.134h-4.956c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248h1.302c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm-69.158 0v-1.134H1.918c.056-.56.28-.882.56-1.218.364-.378.63-.616 1.806-1.358 1.834-1.148 2.534-2.044 2.534-3.304 0-1.596-1.162-2.758-3.01-2.758-1.96 0-3.332 1.176-3.36 3.248H1.75c.056-1.442.742-2.184 1.946-2.184 1.05 0 1.778.588 1.778 1.624 0 .966-.56 1.582-2.128 2.534-1.148.714-1.736 1.232-2.142 1.722-.462.588-.742 1.512-.742 2.422v.406h6.412zm4.611-9.772c-2.156 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998s3.5-1.694 3.5-4.998c0-3.22-1.344-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92zm6.523 8.708h1.288v-9.562h-1.092c-.126 1.19-.728 1.624-2.324 1.708v.91h2.128v6.944zm6.794-9.772c-2.157 0-3.486 1.764-3.486 4.984 0 3.22 1.33 4.998 3.486 4.998 2.155 0 3.5-1.694 3.5-4.998 0-3.22-1.345-4.984-3.5-4.984zm0 1.064c1.428 0 2.114 1.372 2.114 3.92 0 2.492-.686 3.934-2.114 3.934-1.442 0-2.114-1.442-2.114-3.934 0-2.548.672-3.92 2.114-3.92z"
-        />
-        <path
-          fill="#C8C4C5"
-          d="M330.828 265.647h.882v-2.574h1.044c.81 0 1.422-.099 1.791-.315.531-.324.882-.891.882-1.674 0-.738-.297-1.377-.873-1.683-.567-.288-1.197-.288-1.746-.288h-1.98v6.534zm3.69-4.563c0 .486-.216.927-.621 1.134-.207.099-.603.135-1.053.135h-1.134v-2.52h1.134c.45 0 .846.018 1.053.126.405.207.621.639.621 1.125zm1.859-.081v4.644h.828v-2.133c0-1.125.234-1.845 1.233-1.845.162 0 .342.027.549.063v-.729a2.57 2.57 0 00-.522-.054c-.513 0-.945.18-1.305.738v-.684h-.783zm3.311 0v4.644h.828v-4.644h-.828zm0-.99h.828v-.9h-.828v.9zm5.123 4.077c-.144.702-.558 1.035-1.161 1.035-.837 0-1.314-.72-1.314-1.8 0-1.071.459-1.791 1.314-1.791.621 0 .999.333 1.098.99h.864c-.144-1.044-.837-1.656-1.953-1.656-1.395 0-2.187 1.053-2.187 2.457 0 1.467.765 2.457 2.169 2.457 1.134 0 1.782-.558 1.98-1.692h-.81zm3.575 1.692c1.062 0 1.917-.522 2.115-1.521h-.873c-.171.594-.621.864-1.215.864-.756 0-1.323-.621-1.368-1.539h3.465c0-1.719-.657-2.718-2.106-2.718-1.368 0-2.223 1.053-2.223 2.457 0 1.53.783 2.457 2.205 2.457zm-.018-4.248c.666 0 1.17.441 1.26 1.377h-2.565c.117-.936.603-1.377 1.305-1.377zm4.803-.531v4.644h.828v-4.644h-.828zm0-.99h.828v-.9h-.828v.9zm2.873.99h-.765v4.644h.828v-2.43c0-1.116.459-1.656 1.233-1.656.891 0 .99.747.99 1.467v2.619h.828v-2.781c0-.441-.036-.819-.162-1.134-.225-.522-.675-.828-1.476-.828-.468 0-1.053.135-1.476.693v-.594zm-23.785 14.644h.891v-5.805h2.151v-.729h-5.193v.729h2.151v5.805zm4.568-4.059v-2.475h-.828v6.534h.828v-2.43c0-1.116.459-1.656 1.233-1.656.891 0 .99.756.99 1.476v2.61h.828v-2.781c0-.441-.027-.819-.153-1.134-.225-.522-.684-.828-1.485-.828-.522 0-1.08.153-1.413.684zm6.202 3.537c-.846 0-1.377-.729-1.377-1.809 0-1.071.54-1.782 1.386-1.782.855 0 1.404.729 1.404 1.8 0 1.071-.54 1.791-1.413 1.791zm0 .657c1.404 0 2.268-1.008 2.268-2.457 0-1.44-.873-2.457-2.259-2.457-1.377 0-2.241 1.017-2.241 2.457 0 1.449.855 2.457 2.232 2.457zm6.274-.135h.765v-4.644h-.819v2.439c0 1.116-.459 1.647-1.233 1.647-.891 0-.981-.756-.981-1.467v-2.619h-.819v2.7c0 .441 0 .9.153 1.215.243.495.657.828 1.467.828.468 0 1.053-.135 1.467-.693v.594zm1.72-1.494c.009 1.035.738 1.629 2.025 1.629 1.089 0 1.89-.549 1.89-1.386 0-1.152-.972-1.314-1.8-1.476-.612-.117-1.143-.234-1.143-.756 0-.396.369-.648.936-.648.711 0 1.044.297 1.053.909h.819c-.009-1.008-.684-1.557-1.854-1.557-1.098 0-1.818.558-1.818 1.35 0 1.134.918 1.296 1.728 1.44.63.126 1.197.225 1.197.792 0 .45-.36.693-.99.693-.765 0-1.206-.333-1.206-.99h-.837zm8.565 1.494c-.09-.423-.117-.864-.117-1.404v-1.269c0-.936-.081-2.106-1.836-2.106-1.116 0-1.899.612-1.899 1.566h.819c0-.621.432-.945 1.08-.945.711 0 1.026.396 1.026 1.089v.162c-1.152.135-3.096.081-3.096 1.602 0 .819.585 1.404 1.611 1.404.666 0 1.179-.207 1.575-.729.009.315.036.495.063.63h.774zm-.927-2.268v.261c0 1.233-.63 1.503-1.296 1.503-.594 0-.963-.333-.963-.783 0-.945 1.557-.891 2.259-.981zm2.816-2.376h-.765v4.644h.828v-2.43c0-1.116.459-1.656 1.233-1.656.891 0 .99.747.99 1.467v2.619h.828v-2.781c0-.441-.036-.819-.162-1.134-.225-.522-.675-.828-1.476-.828-.468 0-1.053.135-1.476.693v-.594zm7.462 4.644h.792v-6.534h-.828v2.43c-.306-.468-.774-.639-1.404-.639-1.206 0-1.989 1.035-1.989 2.421 0 1.548.765 2.421 1.989 2.421.63 0 1.125-.243 1.44-.711v.612zm-.018-2.322c0 1.08-.522 1.764-1.305 1.764-.765 0-1.251-.684-1.251-1.764 0-1.107.486-1.764 1.251-1.764.783 0 1.305.657 1.305 1.764zm1.764.828c.009 1.035.738 1.629 2.025 1.629 1.089 0 1.89-.549 1.89-1.386 0-1.152-.972-1.314-1.8-1.476-.612-.117-1.143-.234-1.143-.756 0-.396.369-.648.936-.648.711 0 1.044.297 1.053.909h.819c-.009-1.008-.684-1.557-1.854-1.557-1.098 0-1.818.558-1.818 1.35 0 1.134.918 1.296 1.728 1.44.63.126 1.197.225 1.197.792 0 .45-.36.693-.99.693-.765 0-1.206-.333-1.206-.99h-.837z"
-        />
-        <path fill="#000" d="M124.582 283.647v-9h-1.164v9h1.164z" />
-        <path
-          fill="#8D8D8D"
-          d="M3.179 76.21H2v-.946h1.179v.946zm3.536 0H4.357v-.946h2.358v.946zm3.536 0H7.893v-.946h2.358v.946zm3.535 0H11.43v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H29.11v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H36.18v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0H53.86v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.358v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H71.54v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H89.22v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0H96.29v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm2.357 0h-1.178v-.946h1.178v.946zM3.179 208.75H2v-.947h1.179v.947zm3.536 0H4.357v-.947h2.358v.947zm3.536 0H7.893v-.947h2.358v.947zm3.535 0H11.43v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0H29.11v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0H36.18v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0H53.86v-.947h2.358v.947zm3.535 0h-2.357v-.947h2.358v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0H71.54v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0H89.22v-.947h2.357v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0H96.29v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.535 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.535 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.535 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.536 0h-2.358v-.947h2.358v.947zm3.535 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm3.536 0h-2.357v-.947h2.357v.947zm2.357 0h-1.178v-.947h1.178v.947zM3.179 8.757H2V7.81h1.179v.947zm3.536 0H4.357V7.81h2.358v.947zm3.536 0H7.893V7.81h2.358v.947zm3.535 0H11.43V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0H29.11V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0H36.18V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0H53.86V7.81h2.358v.947zm3.535 0h-2.357V7.81h2.358v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0H71.54V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0H89.22V7.81h2.357v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0H96.29V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.535 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.535 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.535 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.536 0h-2.358V7.81h2.358v.947zm3.535 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm3.536 0h-2.357V7.81h2.357v.947zm2.357 0h-1.178V7.81h1.178v.947zM3.179 142.48H2v-.946h1.179v.946zm3.536 0H4.357v-.946h2.358v.946zm3.536 0H7.893v-.946h2.358v.946zm3.535 0H11.43v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H29.11v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H36.18v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0H53.86v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.358v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H71.54v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0H89.22v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0H96.29v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.536 0h-2.358v-.946h2.358v.946zm3.535 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm3.536 0h-2.357v-.946h2.357v.946zm2.357 0h-1.178v-.946h1.178v.946z"
-        />
-        <path
-          fill="#fff"
-          d="M315.982 76.319h-9.597v-1.164h9.597v1.164zm0 132.539h-9.597v-1.163h9.597v1.163zm0-199.992h-9.597V7.702h9.597v1.164zm0 133.77h-9.597v-1.164h9.597v1.164zm0 132.493h-9.597v-1.164h9.597v1.164z"
-        />
-        <path
-          fill="#8D8D8D"
-          d="M1.526 9.463V8.284h.947v1.179h-.947zm0 3.535V10.64h.947v2.357h-.947zm0 3.534v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.536V31.85h.947v2.357h-.947zm0 3.535v-2.357h.947v2.356h-.947zm0 3.535V38.92h.947v2.357h-.947zm0 3.535v-2.357h.947v2.356h-.947zm0 3.534v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.357h-.947zm0 3.536v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 2.356v-1.178h.947v1.178h-.947zM62.661 9.462V8.284h.947v1.178h-.947zm0 3.535V10.64h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V17.71h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V24.78h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V31.85h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V38.92h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V45.99h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V53.06h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V60.13h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V67.2h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V74.27h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V81.34h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V88.41h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V95.48h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V137.9h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V208.6h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 2.357v-1.178h.947v1.178h-.947zM123.003 9.462V8.284h.947v1.178h-.947zm0 3.535V10.64h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V17.71h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V24.78h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V31.85h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V38.92h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V45.99h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V53.06h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V60.13h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V67.2h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V74.27h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V81.34h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V88.41h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V95.48h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V137.9h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V208.6h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 2.357v-1.178h.947v1.178h-.947zM184.139 9.462V8.284h.946v1.178h-.946zm0 3.535V10.64h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V17.71h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V24.78h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V31.85h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V38.92h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V45.99h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V53.06h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V60.13h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V67.2h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V74.27h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V81.34h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V88.41h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V95.48h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V137.9h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535V208.6h.946v2.357h-.946zm0 3.535v-2.357h.946v2.357h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 3.535v-2.356h.946v2.356h-.946zm0 2.357v-1.178h.946v1.178h-.946zM245.275 9.462V8.284h.947v1.178h-.947zm0 3.535V10.64h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V17.71h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V24.78h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V31.85h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V38.92h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V45.99h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V53.06h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V60.13h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V67.2h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V74.27h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V81.34h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V88.41h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V95.48h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V137.9h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535V208.6h.947v2.357h-.947zm0 3.535v-2.357h.947v2.357h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 3.535v-2.356h.947v2.356h-.947zm0 2.357v-1.178h.947v1.178h-.947z"
-        />
-        <path
-          fill="#FFFCFC"
-          d="M185.195 284.014v-9.468h-1.164v9.468h1.164zm-61.135 0v-9.468h-1.164v9.468h1.164zm122.271 0v-9.468h-1.164v9.468h1.164zm60.351-1.184v-8.284h-1.183v8.284h1.183z"
-        />
-        <path
-          fill="#FFFCFC"
-          d="M305.499 276.913V8.283h1.183v268.63h-1.183zm-241.781 7.101v-9.468h-1.163v9.468h1.163zm-62.299 0v-10.651h1.164v10.651H1.419z"
-        />
-        <path fill="#fff" d="M306 275.229H2v-1.164h304v1.164z" />
-        <path
-          fill="#066AE0"
-          d="M301.943 9.59a5 5 0 013.967 5.854l-24.5 127.499a5 5 0 01-9.653.638L246.926 69.09l-26.549 117.512a4.985 4.985 0 01-.677 1.611l-31 48a5.001 5.001 0 01-8.188.303l-29.337-38.784-29.498-31.314-24.64 15.452-29.395 73.487a5 5 0 01-9.284 0l-26.095-65.236-19.892 35.807a5 5 0 11-8.742-4.856l25-45a5 5 0 019.013.571L63 240.037l25.358-63.394a5.003 5.003 0 011.986-2.379l29.499-18.5a5 5 0 016.296.807l32.5 34.5c.124.131.24.269.349.413l25.189 33.3 26.641-41.251 30.305-134.135a5 5 0 019.62-.48l24.455 73.363L296.09 13.556a5 5 0 015.853-3.966z"
-        />
-      </svg>
-    </ResponsiveBox>
-  )
-}
