+++ src/v2/Apps/Consign/Routes/MarketingLanding/Components/PromoSpace.tsx
@@ -1,18 +1,30 @@
-import * as React from "react"
-import { Banner, Box, FullBleed } from "@artsy/palette"
+import { Flex, FullBleed, Text, ArrowRightIcon } from "@artsy/palette"
+import { AppContainer } from "v2/Apps/Components/AppContainer"
+import { HorizontalPadding } from "v2/Apps/Components/HorizontalPadding"
+import { RouterLink } from "v2/System/Router/RouterLink"
 
 export const PromoSpace: React.FC = () => {
   return (
-    <FullBleed>
-      <Banner variant="defaultLight">
-        Gallerist or art dealer?&nbsp;
-        <a href="https://partners.artsy.net" target="_blank">
-          Become a partner
-        </a>
-        <Box as="span" display={["none", "inline"]}>
-          &nbsp;to access the largest global online art marketplace
-        </Box>
-      </Banner>
-    </FullBleed>
+    <RouterLink
+      to="https://partners.artsy.net"
+      textDecoration="none"
+      target="_blank"
+    >
+      <FullBleed backgroundColor="black5">
+        <AppContainer py={1}>
+          <HorizontalPadding>
+            <Flex justifyContent="space-between" alignItems="center">
+              <Text mr={2} variant={["xs", "md"]}>
+                Gallery or art dealer? <u>Become a partner</u> to access the
+                world’s largest online marketplace.
+              </Text>
+              <Flex>
+                <ArrowRightIcon />
+              </Flex>
+            </Flex>
+          </HorizontalPadding>
+        </AppContainer>
+      </FullBleed>
+    </RouterLink>
   )
 }
