+++ src/v2/Apps/Consign/Routes/MarketingLanding/Components/__tests__/FAQ.jest.tsx
@@ -1,29 +1,13 @@
-import { mount } from "enzyme"
-import { MockBoot } from "v2/DevTools"
 import { FAQ } from "../FAQ"
-import { Breakpoint } from "@artsy/palette"
-import { AnalyticsSchema, useTracking } from "v2/System"
 import { ContextModule, OwnerType } from "@artsy/cohesion"
+import { fireEvent, render, screen } from "@testing-library/react"
+import { useTracking, AnalyticsSchema } from "v2/System/Analytics"
 
 jest.mock("v2/System/Analytics/useTracking")
 
-jest.mock("v2/System/Router/useRouter", () => ({
-  useRouter: jest.fn(() => ({
-    match: { params: { id: "1" } },
-  })),
-}))
-
-const trackEvent = useTracking as jest.Mock
+const trackEvent = jest.fn()
 
 describe("FAQ", () => {
-  const getWrapper = (breakpoint = "md") => {
-    return mount(
-      <MockBoot breakpoint={breakpoint as Breakpoint}>
-        <FAQ />
-      </MockBoot>
-    )
-  }
-
   beforeEach(() => {
     ;(useTracking as jest.Mock).mockImplementation(() => {
       return {
@@ -32,21 +16,38 @@ describe("FAQ", () => {
     })
   })
 
-  it("toggles FAQ questions", () => {
-    const wrapper = getWrapper()
-    expect(wrapper.html()).not.toContain(
-      "We offer no upfront fees." // hidden
-    )
-    wrapper.find("Clickable").first().simulate("click")
-    wrapper.update()
-    expect(wrapper.html()).toContain(
-      "We offer no upfront fees." // visible on toggle click
-    )
+  afterEach(() => {
+    jest.clearAllMocks()
   })
 
-  it("tracks a FAQ click with correct params when submission id found", () => {
-    const wrapper = getWrapper()
-    wrapper.find("Clickable").first().simulate("focus")
+  it("toggles FAQ questions", () => {
+    render(<FAQ />)
+
+    const question = screen.getByText("What does it cost to sell with Artsy?")
+    expect(question).toBeInTheDocument()
+    expect(
+      screen.queryByText(/^There are no upfront fees./)
+    ).not.toBeInTheDocument()
+
+    fireEvent.click(question)
+
+    expect(
+      screen.queryByText(/^There are no upfront fees./)
+    ).toBeInTheDocument()
+  })
+
+  it("doesn't track a FAQ click if there is no needs", () => {
+    render(<FAQ />)
+
+    fireEvent.focus(screen.getByText("What does it cost to sell with Artsy?"))
+
+    expect(trackEvent).not.toHaveBeenCalled()
+  })
+
+  it("tracks a FAQ click with correct params", () => {
+    render(<FAQ shouldTrackClickEvent />)
+
+    fireEvent.focus(screen.getByText("What does it cost to sell with Artsy?"))
 
     expect(trackEvent).toHaveBeenCalled()
     expect(trackEvent).toHaveBeenCalledWith({
