+++ src/v2/Apps/Consign/Routes/MarketingLanding/Components/__tests__/SoldRecentlyOnArtsy.jest.tsx
