--- src/v2/Apps/Consign/Routes/MarketingLanding/Utils/__tests__/formatCentsToDollars.jest.tsx
@@ -1,12 +1,0 @@
-import { formatCentsToDollars } from "../formatCentsToDollars"
-
-describe("formatCentsToDollars", () => {
-  it("ignores negative numbers", () => {
-    expect(formatCentsToDollars(-1)).toBe("$0")
-  })
-
-  it("formats inputs correctly", () => {
-    expect(formatCentsToDollars(200)).toBe("$2")
-    expect(formatCentsToDollars(200.013)).toBe("$2")
-  })
-})
