--- src/v2/Apps/Consign/Routes/MarketingLanding/Utils/formatCentsToDollars.ts
@@ -1,17 +1,0 @@
-export function formatCentsToDollars(value: number) {
-  if (value < 0) {
-    value = 0
-  }
-
-  const CENTS_IN_DOLLAR = 100
-
-  return new Intl.NumberFormat("en-US", {
-    // TODO: While artwork price insight stuff is all currently calculated
-    // in dollars, we might need to return to this later.
-    currency: "USD",
-
-    style: "currency",
-  })
-    .format(Math.round(value / CENTS_IN_DOLLAR))
-    .replace(".00", "")
-}
